+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,6 +6,12 @@ Unreleased
 ### Changed
 * Update the filterDate examples to handle an ISO string instead of a moment object.
 
+### Removed
+* Remove minDateTime and maxDateTime props. These props are being renamed to minDate and maxDate.
+
+### Added
+* Add minDate and maxDate props. These props replace minDateTime and maxDateTime for the purpose of ignoring the time portion when determing the min/max range.
+
 3.16.0 - (May 7, 2019)
 ------------------
 ### Fixed
@@ -19,6 +25,7 @@ Unreleased
 ### Changed
 * Moved packages we use solely in doc examples and tests to devDependencies
 * Leverage createSafeDate from the DateUtil.
+
 
 3.14.0 - (April 16, 2019)
 ------------------